+++ be/src/vec/exprs/vruntimefilter_wrapper.h
@@ -77,7 +77,9 @@ public:
     template <typename T>
     static void judge_selectivity(double ignore_threshold, int64_t filter_rows, int64_t input_rows,
                                   T& always_true) {
-        always_true = filter_rows / (input_rows * 1.0L) < ignore_threshold;
+        always_true = static_cast<long double>(filter_rows) /
+                              (static_cast<long double>(input_rows) * 1.0L) <
+                      ignore_threshold;
     }
 
     bool is_rf_wrapper() const override { return true; }
